chore: reorganize option models and functions
