+++ models/options/GlitchOptionSet.py
@@ -1,0 +1,155 @@
+
+class GlitchOptionSet():
+    def __init__(self):
+        self.prologue_gel_early = False
+        self.reverse_goomba_king_bridge = False
+        self.goomba_village_entry_fence_clip = False
+        self.goomba_village_npc_lure_exit = False
+        self.hammerless_jr_playground_laki = False
+        self.goomba_village_laki_exit = False
+        self.prologue_sushie_glitch_ksj = False
+        self.prologue_sushie_glitch_ultra_boots_laki = False
+
+        self.odd_key_early = False
+        self.blue_house_skip = False
+        self.blue_house_skip_laki = False
+        self.blue_house_skip_toad_lure = False
+        self.bowless_toy_box_hammer = False
+        self.bowless_toy_box_hammerless_lure = False
+        self.early_storeroom_parakarry = False
+        self.early_storeroom_hammer = False
+        self.early_storeroom_hammerless_lure = False
+        self.whale_early = False
+        self.sushiesless_toad_town_star_piece = False
+        self.toad_town_sushie_glitch = False
+
+        self.clippy_boots_stone_block_skip = False
+        self.clippy_boots_metal_block_skip = False
+        self.island_pipe_blooper_skip = False
+        self.parakarryless_sewer_star_piece = False
+        self.sewer_blocks_without_ultra_boots = False
+        self.first_block_to_shiver_city_without_super_boots = False
+        self.blocks_to_shiver_city_kooper_shell_item_throw = False
+        self.sewer_yellow_block_with_ultra_boots = False
+        self.jumpless_sewer_shooting_star = False
+
+        self.kooperless_pleasant_path_star_piece = False
+        self.hammerless_pleasant_path_bridge_ultra_boots_parakarry = False
+        self.invisible_bridge_clip_lzs = False
+        self.invisible_bridge_clip_laki = False
+        self.kooperless_pleasant_path_thunderbolt = False
+
+        self.bombetteless_kbf_fp_plus_lzs = False
+        self.bombetteless_kbf_fp_plus_laki = False
+        self.laki_jailbreak = False
+        self.bombetteless_right_fortress_jail_key = False
+        self.water_staircase_skip = False
+
+        self.mt_rugged_quake_hammer_and_letter_with_laki = False
+        self.parakarryless_mt_rugged_seed = False
+        self.buzzar_gap_skip_clippy = False
+        self.mt_rugged_coins_with_kooper = False
+        self.mt_rugged_station_jumpless_climb_bombette = False
+        self.mt_rugged_station_jumpless_climb_laki = False
+        self.jumpless_mt_rugged_train_platform_parakarry = False
+        self.parakarryless_mt_rugged_star_piece = False
+        self.desert_brick_block_item_with_parakarry = False
+        self.early_ruins_laki_jump = False
+        self.early_ruins_ultra_boots = False
+
+        self.artifact_jump_laki = False
+        self.artifact_jump_ultra_boots = False
+        self.ruins_key_laki_jump = False
+        self.parakarryless_second_sand_room_ultra_boots = False
+        self.parakarryless_second_sand_room_normal_boots = False
+        self.parakarryless_super_hammer_room_ultra_boots = False
+        self.parakarryless_super_hammer_room_normal_boots = False
+        self.ruins_locks_skip_clippy = False
+
+        self.record_skip_no_bombette_push = False
+        self.record_skip_bombette_push = False
+        self.boos_portrait_with_kooper = False
+        self.boos_portrait_with_laki = False
+        self.jumpless_mansion_entry = False
+
+        self.gusty_gulch_gate_skip_lzs = False
+        self.gusty_gulch_gate_skip_laki = False
+        self.kooperless_gusty_gulch_dizzy_dial_jump = False
+        self.kooperless_gusty_gulch_dizzy_dial_laki = False
+        self.kooperless_gusty_gulch_dizzy_dial_parakarry = False
+        self.gusty_gulch_gap_skip = False
+
+        self.bowless_tubbas_castle = False
+        self.tubbas_table_laki_jump_clock = False
+        self.tubbas_table_ultra_boots = False
+        self.tubbas_table_laki_jump_study = False
+        self.tubbas_castle_super_boots_skip = False
+        self.parakarryless_mega_rush = False
+
+        self.parakarryless_blue_building_star_piece = False
+        self.gourmet_guy_skip_jump = False
+        self.gourmet_guy_skip_laki = False
+        self.gourmet_guy_skip_parakarry = False
+        self.bowless_green_station = False
+        self.kooperless_red_station_shooting_star = False
+        self.gearless_red_station_shooting_star = False
+        self.parakarryless_blue_block_city_gap = False
+        self.blue_switch_skip_laki = False
+        self.blue_switch_skip_ultra_boots = False
+        self.red_barricade_skip = False
+        self.hammerless_blue_station_laki = False
+        self.hammerless_pink_station_laki = False
+
+        self.raph_skip_english = False
+        self.raph_skip_parakarry = False
+        self.ch5_sushie_glitch = False
+        self.sushieless_jungle_starpiece_and_letter = False
+        self.jumpless_deep_jungle_laki = False
+        self.kooperless_lavalava_pow_block_parakarry = False
+        self.kooperless_lavalava_pow_block_super_boots = False
+        self.jumpless_lavalava_pow_block = False
+        self.ultra_hammer_skip = False
+        self.ultra_hammer_skip_laki = False
+        self.ultra_hammer_skip_sushie = False
+        self.flarakarry = False
+        self.parakarryless_flarakarry_bombette = False
+        self.parakarryless_flarakarry_laki = False
+        self.volcano_sushie_glitch = False
+
+        self.early_laki_lzs = False
+        self.early_laki_bombette_push = False
+        self.bombetteless_mega_smash = False
+        self.sun_tower_skip = False
+        self.yellow_berry_gate_skip_lzs = False
+        self.yellow_berry_gate_skip_laki = False
+        self.yellow_berry_gate_skip_bombette_push = False
+        self.red_berry_gate_skip_bombette_push = False
+        self.red_berry_gate_skip_laki = False
+        self.blue_berry_gate_skip_bombette_push = False
+        self.blue_berry_gate_skip_laki = False
+        self.bubble_berry_tree_early_laki_jump = False
+        self.bubble_berry_tree_early_ultra_boots = False
+
+        self.murder_solved_early_laki = False
+        self.murder_solved_early_bombette_push = False
+        self.ch7_sushie_glitch = False
+        self.star_stone_with_ch7_sushie_glitch = False
+        self.shiver_mountain_hidden_block_without_ultra_boots_laki = False
+        self.shiver_mountain_hidden_block_without_ultra_boots_no_laki = False
+        self.snowmen_skip_laki = False
+        self.shiver_mountain_switch_skip = False
+        self.sushieless_warehouse_key_bombette = False
+        self.sushieless_warehouse_key_kooper = False
+
+        self.mirror_clip = False
+
+        self.bowless_bowsers_castle_basement = False
+        self.fast_flood_room_kooper = False
+        self.fast_flood_room_bombette_ultra_boots = False
+        self.bombetteless_bowsers_castle_basement = False
+
+        self.break_yellow_blocks_with_super_boots = False
+        self.break_stone_blocks_with_ultra_boots = False
+        self.knows_hidden_blocks = False
+        self.knows_puzzle_solutions = False
+        self.reach_high_blocks_with_super_boots = False
