+++ models/options/PaletteOptionSet.py
@@ -1,0 +1,27 @@
+class PaletteOptionSet():
+    def __init__(self):
+        DEFAULT_PALETTE = 0
+
+        self.mario_setting = DEFAULT_PALETTE
+        self.mario_sprite = DEFAULT_PALETTE
+        self.goombario_setting = DEFAULT_PALETTE
+        self.goombario_sprite = DEFAULT_PALETTE
+        self.kooper_setting = DEFAULT_PALETTE
+        self.kooper_sprite = DEFAULT_PALETTE
+        self.bombette_setting = DEFAULT_PALETTE
+        self.bombette_sprite = DEFAULT_PALETTE
+        self.parakarry_setting = DEFAULT_PALETTE
+        self.parakarry_sprite = DEFAULT_PALETTE
+        self.bow_setting = DEFAULT_PALETTE
+        self.bow_sprite = DEFAULT_PALETTE
+        self.watt_setting = DEFAULT_PALETTE
+        self.watt_sprite = DEFAULT_PALETTE
+        self.sushie_setting = DEFAULT_PALETTE
+        self.sushie_sprite = DEFAULT_PALETTE
+        self.lakilester_setting = DEFAULT_PALETTE
+        self.lakilester_sprite = DEFAULT_PALETTE
+
+        self.bosses_setting = DEFAULT_PALETTE
+        self.enemies_setting = DEFAULT_PALETTE
+        self.npc_setting = DEFAULT_PALETTE
+        self.hammer_setting = DEFAULT_PALETTE
