+++ models/options/option_utility.py
@@ -1,0 +1,19 @@
+from db.option import Option
+
+
+def get_option_keyvalue_dict(option_str: str):
+    key = Option.get(Option.name == option_str).get_key()
+    value = Option.get(Option.name == option_str).value
+
+    return {"key": key, "value": value}
+
+
+def populate_keys(data:dict):
+    for option_str in data.copy():
+        if Option.get_or_none(Option.name == option_str) is not None:
+            data[option_str] = {
+                "key": Option.get(Option.name == option_str).get_key(),
+                "value": data[option_str]
+            }
+        else:
+            data[option_str] = {"value": data[option_str]}
